+++ peek_plugin_diagram/_private/storage/branch/BranchIndexEncodedChunk.py
@@ -1,6 +1,6 @@
 import logging
 
-from sqlalchemy import Column, LargeBinary
+from sqlalchemy import Column, LargeBinary, BigInteger
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer, String
 from sqlalchemy.orm import relationship
@@ -22,7 +22,7 @@ class BranchIndexEncodedChunk(Tuple, DeclarativeBase,
     __tablename__ = 'BranchIndexEncodedChunk'
     __tupleType__ = diagramTuplePrefix + 'BranchIndexEncodedChunkTable'
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
 
     modelSetId = Column(Integer,
                         ForeignKey('ModelSet.id', ondelete='CASCADE'),
